put linearsolve in an extension
